+++ route.py
@@ -1,10 +1,19 @@
 import math
 import numpy as np
+from collections import defaultdict
 from pathfinding.core.diagonal_movement import DiagonalMovement
 from pathfinding.core.grid import Grid
 from pathfinding.finder.a_star import AStarFinder
 from pathfinding.finder.breadth_first import BreadthFirstFinder
-# from debug import show_grid, plot_debug_gerber
+from pathfinding3d.core.diagonal_movement import DiagonalMovement as DiagonalMovement3D
+from pathfinding3d.core.grid import Grid as Grid3D
+from pathfinding3d.finder.a_star import AStarFinder as AStarFinder3D
+
+from manipulate import consolidate_segments, merge_overlapping_segments
+
+BLOCKED_CELL = 0
+FREE_CELL = 1
+TUNNEL_CELL = 20
 
 def to_grid_indices(x, y, center_x, center_y, resolution):
     """
@@ -23,6 +32,24 @@ def to_grid_indices(x, y, center_x, center_y, resolution):
     row = int(center_y - (y / resolution))
     col = int(center_x + (x / resolution))
     return row, col
+
+def invert_layer_mapping(layer_mapping):
+    """
+    Build a dict that allows quick lookup of which layer each net is on.
+    Example result:
+      {
+        "JD_PWR": "F_Cu.gtl",
+        "JD_DATA": "F_Cu.gtl",
+        "JD_GND": "B_Cu.gbl",
+      }
+    """
+    net_to_layer = {}
+    for layer_file, info in layer_mapping.items():
+        for net in info["nets"]:
+            # Guard against empty or None in "nets" if needed.
+            if net:
+                net_to_layer[net] = layer_file
+    return net_to_layer
 
 def create_grid(dimensions, keep_out_zones, resolution):
     """
@@ -49,7 +76,7 @@ def create_grid(dimensions, keep_out_zones, resolution):
     center_x, center_y = grid_width // 2, grid_height // 2
     
     # Initialize grid with all ones (free cells)
-    grid = np.ones((grid_height, grid_width), dtype=int)
+    grid = np.full((grid_height, grid_width), FREE_CELL, dtype=int)
     
     # Mark keep-out zones in the grid
     for zone in keep_out_zones:
@@ -66,7 +93,7 @@ def create_grid(dimensions, keep_out_zones, resolution):
         y2 = max(0, min(grid_height - 1, y2))
         
         # Mark cells in the rectangle as blocked
-        grid[y1+1:y2, x1+1:x2] = 0 # IDK why, but adding +1 made the grid work correctly (assuming it's an index thing)
+        grid[y1+1:y2, x1+1:x2] = BLOCKED_CELL 
 
     return grid
 
@@ -114,9 +141,9 @@ def apply_socket_keep_out_zones(grid, socket_locations, current_net, resolution,
                     # Check if the keep-out zone for the socket is within the grid boundaries
                     if 0 <= xi < grid.shape[1] and 0 <= yi < grid.shape[0]:
                         if net != current_net:
-                            temp_grid[yi, xi] = 0 # Mark this position as blocked
+                            temp_grid[yi, xi] = BLOCKED_CELL # Mark this position as blocked
                         else:
-                            temp_grid[yi, xi] = 1 # Mark this position as free
+                            temp_grid[yi, xi] = FREE_CELL # Mark this position as free
 
     # For debugging: plot the rectangles on debug.gbr
     # plot_debug_gerber(rectangles, output_file="debug_keepouts.gbr")
@@ -178,138 +205,132 @@ class UnionFind:
                 if self.rank[root1] == self.rank[root2]:
                     self.rank[root2] += 1
 
-def route_sockets(grid, socket_locations, layer_mapping, resolution, algorithm, allow_diagonal_traces, debug=False):
-    """
-    Performs routing for each Gerber Socket on the same net, based on the grid with the keep-out zones. Also
-    applies additional keep-out zones for the socket on other nets to avoid shorts with vias. The Union-Find 
-    (also called Disjoint Set Union or DSU) is used to determine if two sockets belong to the same net, and
-    if they are already connected. For each pair of sockets, if they are connected, no need to route again, 
-    and if not, perform pathfinding - and if successful - union them together.
-
-    Parameters:
-        grid (numpy.array): The grid on which to perform the routing, with obstacles marked.
-                            socket_locations (dict): A dictionary of socket locations grouped by net names.
-        resolution (float): The resolution of the grid in units per grid cell.
-        algorithm (str): Optional, the pathfinding algorithm to use, either 'a_star' or 'breadth_first'.
-        allow_diagonal_traces (bool): Optional, whether to allow diagonal movement in pathfinding. 
-        
-    Returns:
-        dict: A dictionary of nets and segments, where each key is a net name and the value is a list of tuples of line segments
-              ((start_x, start_y), (end_x, end_y)) in the Gerber coordinates.
-    """    
-
-    paths = {}
+def print_debug_grid(matrix, path=None, start=None, end=None):
+    debug_grid = Grid(matrix=matrix)
+    print(debug_grid.grid_str(path, start, end, show_weight=True))
+    
+def route_sockets(grid, socket_locations, configuration):
+    
+    # Unpack the JSON configuration
+    algorithm = configuration['algorithm']
+    allow_diagonal_traces = configuration['allow_diagonal_traces']
+    resolution = configuration['resolution']
+    layer_mapping = configuration['layer_mapping']
+    
+    paths = {} # {net: list of grid indeces (x, y) for path}
+    previous_paths = defaultdict(list) # {layer_filename: list of path-indicies}
+    
+    # Calculate the distances between the sockets for each net
     net_distances = calculate_net_distances(socket_locations, resolution)
     
     # Center coordinates of the grid
-    grid_width = grid.shape[1]
     grid_height = grid.shape[0]
+    grid_width = grid.shape[1]
     center_x, center_y = grid_width // 2, grid_height // 2
+    
+    # Net to layer mapping for quicker lookups
+    net_to_layer_mapping = invert_layer_mapping(layer_mapping)
 
     for net, distances in net_distances.items():
-        temp_grid = apply_socket_keep_out_zones(grid, socket_locations, net, resolution)
-        uf = UnionFind([tuple(loc) for loc in socket_locations[net]])  # Using tuples as UnionFind elements
+        # Identify the layer for the current net (if any)
+        current_layer = net_to_layer_mapping.get(net, None)
+        other_nets_on_layer = False
         
-        print(f"🔵 Applied temporary grid for sockets on net {net}")
+        # Apply the socket keep-out zones - to either expose or block sockets from other nets
+        current_matrix = apply_socket_keep_out_zones(grid, socket_locations, net, resolution)
+        tunnel_matrix = np.copy(current_matrix) # Make a copy for later tunnelling
+        
+        # Block previously routed paths from all nets on the current layer
+        if previous_paths[current_layer]:
+            other_nets_on_layer = True
+            for path_indexes in previous_paths[current_layer]:
+                for (x_index, y_index, z_index) in path_indexes:
+                    if 0 <= x_index < grid_height and 0 <= y_index < grid_width:
+                        current_matrix[y_index, x_index] = BLOCKED_CELL
+        
+        # A Union-Find for sockets on the current net, so that only routing of new lines will be made if essential
+        uf = UnionFind([tuple(loc) for loc in socket_locations[net]]) 
 
+        # Iterate over the distances between the sockets, sorted in ascending order
         for dist, loc1, loc2 in distances:
+            
+            tunnels_needed = False
+            
             # Check if the two locations are already connected in the union-find structure
-            if uf.find(tuple(loc1)) != uf.find(tuple(loc2)):
+            if uf.find(tuple(loc1)) == uf.find(tuple(loc2)):
+                continue  # already connected, skip
                 
-                # Convert the start and end coordinates to grid indices
-                start_index = to_grid_indices(-loc1[1], -loc1[0], center_x, center_y, resolution)
-                end_index = to_grid_indices(-loc2[1], -loc2[0], center_x, center_y, resolution)
+            # Convert the start and end coordinates to grid indices
+            start_index = to_grid_indices(-loc1[1], -loc1[0], center_x, center_y, resolution)
+            end_index = to_grid_indices(-loc2[1], -loc2[0], center_x, center_y, resolution)
+            
+            # Create a 2D grid for the pathfinding
+            net_grid = Grid(matrix=current_matrix, grid_id=0)
+
+            if algorithm == "breadth_first":
+                finder = BreadthFirstFinder(diagonal_movement=(DiagonalMovement.always if allow_diagonal_traces else DiagonalMovement.never))
+            if algorithm == "a_star":
+                finder = AStarFinder(diagonal_movement=(DiagonalMovement.always if allow_diagonal_traces else DiagonalMovement.never))
+            
+            start = net_grid.node(start_index[0], start_index[1])
+            end = net_grid.node(end_index[0], end_index[1])
+            
+            path, runs = finder.find_path(start, end, net_grid)
+            print(f"🔵 Pathfinding runs: {runs}")
+            
+            if not path and other_nets_on_layer:
                 
-                # Pathfinding setup
-                path_grid = Grid(matrix=temp_grid)
+                print(f"🔵 Making a 3D grid to accomondate routing for {net} on layer {current_layer}")
                 
-                if algorithm == "breadth_first":
-                    finder = BreadthFirstFinder(diagonal_movement=(DiagonalMovement.always if allow_diagonal_traces else DiagonalMovement.never))
-                if algorithm == "a_star":
-                    finder = AStarFinder(diagonal_movement=(DiagonalMovement.always if allow_diagonal_traces else DiagonalMovement.never))
+                # Set the weights for the tunnelling matrix
+                tunnel_matrix[tunnel_matrix == FREE_CELL] = TUNNEL_CELL
                 
-                start = path_grid.node(start_index[0], start_index[1])
-                end = path_grid.node(end_index[0], end_index[1])
+                # print_debug_grid(current_matrix, start=start_index, end=end_index)
+                # print_debug_grid(tunnel_matrix, start=start_index, end=end_index)
+                            
+                # Create a 3D matrix for the tunneling, transpose to match the grid indices
+                combined_matrix = np.stack((tunnel_matrix.T, current_matrix.T), axis=2)
                 
-                path, runs = finder.find_path(start, end, path_grid)
+                # Create a 3D grid for the pathfinding
+                tunnel_grid = Grid3D(matrix=combined_matrix)
+                
+                # Only use A* for 3D pathfinding, as it takes the direction and cost into the heuristic calculation
+                # Routing diagonally does not work in the current implementation 
+                finder = AStarFinder3D(diagonal_movement=(DiagonalMovement3D.never))
+                    
+                start = tunnel_grid.node(start_index[0], start_index[1], 0)
+                end = tunnel_grid.node(end_index[0], end_index[1], 0)
+                
+                path, runs = finder.find_path(start, end, tunnel_grid)
                 print(f"🔵 Pathfinding runs: {runs}")
                 
-                if path:
-                    print(f"🟢 Found path for net {net} between {loc1} and {loc2}")
-
-                    # Convert GridNodes -> (row, col)
-                    path_tuples = [(node.x, node.y) for node in path]
-                    
-                    # Add it to our dictionary
-                    paths.setdefault(net, []).append(path_tuples)
-                    
-                    # Union-Find logic for an optimised routing strategy
-                    uf.union(tuple(loc1), tuple(loc2))
+                tunnels_needed = True
+                
+            if path:
+                print(f"🟢 Found path for net {net} between {loc1} and {loc2}")
+                
+                if tunnels_needed:
+                    path_tuples = [(node.x, node.y, node.z) for node in path]
+                    # tunnel_grid.visualize(path=path, start=start, end=end, visualize_weight=True, save_html=True, save_to=f"./path_visualization_{net}.html")
                 else:
-                    print(f"🔴 No path found for net {net} between {loc1} and {loc2}")
-    
+                    path_tuples = [(node.x, node.y, 1) for node in path]
+                
+                paths.setdefault(net, []).append(path_tuples) # Add it to our dictionary of paths
+                uf.union(tuple(loc1), tuple(loc2)) # Union-Find logic for an optimised routing strategy         
+                previous_paths[current_layer].append(path_tuples) # Add path to list of previous paths
+                
+            else:
+                print(f"🔴 No path found for net {net} between {loc1} and {loc2}")
+                
     # Get the centers of the grid for translating the indices
     center_x = grid.shape[0] // 2
     center_y = grid.shape[1] // 2
     
     # Consolidate the the list of grid indices into line segments on the coordinate plane
     segments = consolidate_segments(paths, resolution, center_x, center_y)
+    
+    # Merge overlapping segments 
+    # segments = merge_overlapping_segments(segments)
 
     return segments
                 
-def consolidate_segments(routes, resolution, center_x, center_y):
-    """
-    Consolidates contiguous/overlaping line segments into longer lines. This translates the indices from the grid to 
-    line ednpoints that can be used for creating traces using `gerber-writer`.
-
-    Parameters:
-        routes (dict): The routing paths indexed by net names, each path is a list of grid indices.
-        resolution (float): The grid resolution.
-        center_x (int): X-coordinate of the grid center.
-        center_y (int): Y-coordinate of the grid center.
-
-    Returns:
-        dict: A dictionary where each key is a net name and the value is a list of line segments
-              (start point, end point) in real-world coordinates.
-    """
-    consolidated_routes = {}
-
-    for net, paths in routes.items():
-        consolidated_paths = []
-        for path in paths:
-            if len(path) < 2:
-                continue
-            # Start the new segment
-            current_segment_start = path[0]
-            current_direction = None
-
-            for i in range(1, len(path)):
-                # Calculate direction
-                previous = path[i - 1]
-                current = path[i]
-                direction = (current[0] - previous[0], current[1] - previous[1])
-
-                if current_direction is None:
-                    current_direction = direction
-                elif direction != current_direction:
-                    # Finish the current segment
-                    consolidated_paths.append((current_segment_start, previous))
-                    
-                    # Start a new segment
-                    current_segment_start = previous
-                    current_direction = direction
-
-                # Extend the segment
-                if i == len(path) - 1:
-                    consolidated_paths.append((current_segment_start, current))
-
-        # Convert grid indices to real-world coordinates
-        real_world_paths = []
-        for start, end in consolidated_paths:
-            real_start = ((start[0] - center_x) * resolution, (center_y - start[1]) * resolution)
-            real_end = ((end[0] - center_x) * resolution, (center_y - end[1]) * resolution) 
-            real_world_paths.append((real_start, real_end))
-
-        consolidated_routes[net] = real_world_paths
-
-    return consolidated_routes