+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - [ ] Check the correctness and shape of the JSON file, and validity of the file names, etc.
 - [ ] Implement error handling
 - [ ] Implement a server endpoint
+- [ ] Add a flag for flooding an entire layer
+- [ ] Merging overlapping tunnel segments causes a via to disappear
+- [ ] Allow for diagonals for tunnels?
+
 
 **Issues**
 
@@ -41,3 +45,8 @@
 - [x] Improved layer mappings, and set it up so that keys are layers, and it only takes a list of nets
 - [x] Fixed the way layer mappings are passed onto the generate function, the EMPTY net is not necessary anymore
 - [x] Began to implement a new method to perform intersections - working on a new way using elevators layers of 2D grid = 3D grid (worlds)
+
+**Tue 21, Jan 2025** 
+
+Write here tomorrow
+